seperate test for evm circuit
